mysite/requirements.txt
Django==3.2.13
django-braces
django-cors-headers
django-environ
django-gunicorn
django-oauth-toolkit
django-rest-framework-social-oauth2
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
markdown
django-filter
mssql-django
pyodbc
django-user-agents


